# gen/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8080
paths:
...
  /users/{id}:
    delete:
      operationId: deleteUser
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - user-controller
      x-accepts: application/json
    get:
      operationId: getUserById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - user-controller
      x-accepts: '*/*'
    put:
      operationId: updateUser
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - user-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /admin-page/users/{id}:
    delete:
      operationId: deleteUser_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - admin-page-controller
      x-accepts: application/json
    get:
      operationId: getUserById_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
    put:
      operationId: updateUser_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /users/{userId}/favorites/{advertisementId}:
    delete:
      operationId: removeFromFavorites
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: advertisementId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - user-controller
      x-accepts: '*/*'
    post:
      operationId: addToFavorites
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: advertisementId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - user-controller
      x-accepts: '*/*'
  /users/register:
    post:
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - user-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /users/refresh-token:
    post:
      operationId: refreshToken
      parameters:
      - explode: true
        in: cookie
        name: refreshToken
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - user-controller
      x-accepts: '*/*'
...
  /users/login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
      tags:
      - user-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /user-page/favorites/{advertisementId}:
    delete:
      operationId: removeFromFavorites_1
      parameters:
      - explode: false
        in: path
        name: advertisementId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - user-page-controller
      x-accepts: '*/*'
    post:
      operationId: addToFavorites_1
      parameters:
      - explode: false
        in: path
        name: advertisementId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - user-page-controller
      x-accepts: '*/*'
...
  /admin-page/users/{userId}/favorites/{advertisementId}:
    delete:
      operationId: removeFromFavorites_2
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: advertisementId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
    post:
      operationId: addToFavorites_2
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: advertisementId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
  /admin-page/users/register:
    post:
      operationId: createUser_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /users/{userId}/favorites:
    get:
      operationId: getFavorites
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  format: int64
                  type: integer
                type: array
          description: OK
      tags:
      - user-controller
      x-accepts: '*/*'
...
  /admin-page/vehicles/user/{userId}:
    get:
      operationId: getVehiclesByUser
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/VehicleDto'
                type: array
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
...
  /admin-page/users/{userId}/favorites:
    get:
      operationId: getFavorites_2
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  format: int64
                  type: integer
                type: array
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
...
  /admin-page/advertisements/user/{userId}:
    get:
      operationId: getAdvertisementsByUser
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/AdvertisementDto'
                type: array
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
...
components:
...
  schemas:
    VehicleDto:
      example:
        color: color
        year: 6
        description: description
        seats: 2
        hasWarranty: true
        advertisementId: 7
        condition: NEW
        transmission: MANUAL
        fuelType: PETROL
        price: 1.4658129805029452
        model: model
        vin: vin
        id: 0
        power: 0
        brand: brand
        mileage: 5
      properties:
        id:
          format: int64
          type: integer
        brand:
          maxLength: 50
          minLength: 1
          type: string
        model:
          maxLength: 50
          minLength: 1
          type: string
        year:
          format: int32
          type: integer
        price:
          format: double
          type: number
        fuelType:
          enum:
          - PETROL
          - DIESEL
          - LPG
          - ELECTRIC
          - HYBRID
          type: string
        transmission:
          enum:
          - MANUAL
          - AUTOMATIC
          type: string
        color:
          maxLength: 20
          minLength: 1
          type: string
        mileage:
          format: int32
          type: integer
        power:
          format: int32
          minimum: 0
          type: integer
        seats:
          format: int32
          maximum: 8
          minimum: 1
          type: integer
        description:
          maxLength: 2000
          minLength: 0
          type: string
        vin:
          maxLength: 17
          minLength: 17
          type: string
        condition:
          enum:
          - NEW
          - USED
          - DAMAGED
          type: string
        hasWarranty:
          type: boolean
        advertisementId:
          format: int64
          type: integer
      required:
      - advertisementId
      - brand
      - color
      - condition
      - fuelType
      - hasWarranty
      - model
      - price
      - transmission
      - vin
      - year
      type: object
    UserDto:
      example:
        password: password
        role: ADMIN
        id: 0
        email: email
        username: username
      properties:
        id:
          format: int64
          type: integer
        username:
          maxLength: 20
          minLength: 3
          type: string
        email:
          type: string
        password:
          maxLength: 2147483647
          minLength: 6
          type: string
        role:
          enum:
          - ADMIN
          - USER
          - GUEST
          type: string
      required:
      - email
      - password
      - username
      type: object
    AdvertisementDto:
      example:
        id: 0
        title: title
        userId: 6
      properties:
        id:
          format: int64
          type: integer
        title:
          maxLength: 100
          minLength: 10
          type: string
        userId:
          format: int64
          type: integer
      required:
      - title
      type: object
...